requirements.txt
django>=4.0.0
djangorestframework
djangorestframework-simplejwt
pyyaml
django-cors-headers
requests
algoliasearch-django>=2.0,<3.0
python-decouple
